+++ src/main/kotlin/net/botlify/anticaptchacom/enums/TaskStatus.kt
@@ -1,17 +1,8 @@
 package net.botlify.anticaptchacom.enums
 
-enum class TaskStatus(val value: String) {
+import com.fasterxml.jackson.annotation.JsonValue
+
+enum class TaskStatus(@get:JsonValue val value: String) {
     PROCESSING("processing"),
     READY("ready");
-
-    companion object {
-        fun fromValue(value: String): TaskStatus? {
-            return when (value) {
-                "processing" -> PROCESSING
-                "ready" -> READY
-                else -> null
-            }
-        }
-    }
-
 }