+++ src/main/kotlin/net/botlify/anticaptchacom/supplier/utils/StartTask.kt
@@ -2,8 +2,8 @@ package net.botlify.anticaptchacom.supplier.utils
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
-import net.botlify.anticaptchacom.API_URL
 import net.botlify.anticaptchacom.AntiCaptchaComClient
+import org.apache.logging.log4j.LogManager
 
 typealias TaskId = Int
 
@@ -31,14 +31,22 @@ internal class StartTask<T>(
     private val task: T,
 ) {
 
+    companion object {
+        private val log = LogManager.getLogger(StartTask::class.java)
+    }
+
     fun startTask(): TaskId {
         val mapper = jacksonObjectMapper()
         // Send request
+        log.trace("Send request to create task...")
         val request: StartTaskRequest<T> = StartTaskRequest(api.config.apiKey, task, api.config.softId)
+        log.trace("Object to send to create task: {}", request)
         val requestJson = mapper.writeValueAsString(request)
-        val responseString = api.httpRequester.sendPost("$API_URL/createTask", requestJson)
+        log.trace("Json to send to create task: {}", requestJson)
+        val responseString = api.httpRequester.sendPost("/createTask", requestJson)
         // Parse response
-        val response = mapper.readValue<StartTaskResponse>(responseString, StartTaskResponse::class.java)
+        val response = mapper.readValue(responseString, StartTaskResponse::class.java)
+        log.trace("Response of create task: {}", response)
         return response.taskId;
     }