+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 Feature:
-- First version with solving of RECAPTCHA_V2 only.
+- Migrate to Kotlin
+- Add support for Cloudflare Turnstile
+
+Removed:
+- Support for other captcha types than Cloudflare Turnstile